socialplace_service/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
Flask==2.3.3
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycodestyle==2.11.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
